New `types/categorical.ts` to separate q obj; updated `index.ts` & `termdb.ts` to match; incorporated new type into `handlers/categorical.ts`
