+++ server/shared/types/index.ts
@@ -4,3 +4,4 @@ export * from './termsetting'
 export * from './termdb'
 export * from './filter'
 export * from './vocab'
+export * from './categorical'
